Refactored callback query handlers into multiple classes.
